hookstate: get rid of private new() functions.
