--- overlord/hookstate/hooksetup.go
@@ -1,35 +1,0 @@
-// -*- Mode: Go; indent-tabs-mode: t -*-
-
-/*
- * Copyright (C) 2016 Canonical Ltd
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as
- * published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package hookstate
-
-import (
-	"github.com/snapcore/snapd/snap"
-)
-
-// hookSetup is a reference to a hook within a specific snap.
-type hookSetup struct {
-	Snap     string        `json:"snap"`
-	Revision snap.Revision `json:"revision"`
-	Hook     string        `json:"hook"`
-}
-
-func newHookSetup(snap string, revision snap.Revision, hook string) hookSetup {
-	return hookSetup{Snap: snap, Revision: revision, Hook: hook}
-}
