+++ overlord/hookstate/repository_test.go
@@ -21,15 +21,12 @@ package hookstate
 
 import (
 	"regexp"
-	"testing"
 
 	. "gopkg.in/check.v1"
 
 	"github.com/snapcore/snapd/overlord/state"
 	"github.com/snapcore/snapd/snap"
 )
-
-func TestRepository(t *testing.T) { TestingT(t) }
 
 type repositorySuite struct{}
 
@@ -50,8 +47,8 @@ func (s *repositorySuite) TestAddHandlerGenerator(c *C) {
 	state := state.New(nil)
 	state.Lock()
 	task := state.NewTask("test-task", "my test task")
-	hookSetup := newHookSetup("test-snap", snap.R(1), "test-hook")
-	context := newContext(task, hookSetup)
+	setup := hookSetup{Snap: "test-snap", Revision: snap.R(1), Hook: "test-hook"}
+	context := &Context{task: task, setup: setup}
 	state.Unlock()
 
 	c.Assert(context, NotNil)
